requirements.txt
requests
pandas
pyopenssl